fix ios link error for apps consuming `.nupkg`
Xamarin has limited support for `.xcframework` files, they do not have a
built in working solution for NuGet packages yet.
This works around the issue by including `OneSignal.xcframework` as-is
in the `.nupkg`. It also includes a `.target` file which is setting to
add the `OneSignal.xcframework` as a native reference.
See the comment in the `Com.OneSignal.nuspec` file for more details.

The items under `<NativeReference>` in `Com.OneSignal.targets` is from
`OneSignal.iOS.Binding.csproj` in this repo.

This was added based on the recommendations from:
   - https://github.com/xamarin/xamarin-macios/issues/10819#issuecomment-815059416