link sample NSE target project
This was removed in c055282b29c190a08feff24e6a9814d2132df966
due to some build error at the time.
It could have been related to a signing issue with
it naming being com.onesignal.example.OneSignalNotification.
I did encounter a issue with it not copying the app to the device so
I renamed this to
com.onesignal.example.OneSignalNotificationServiceExtensionXamarin
and provision this name on OneSignal's Apple account.

The issue with
com.onesignal.example.OneSignalNotificationServiceExtension
is that OneSignal's Apple account does not own this bundle id.