Fix .targets file being skipped and its path
Without these changes the consuming project will get iOS link errors
as the .xcframework won't be copied in.

The content path was updated in Com.OneSignal.targets but was missed
in the .nuspec file.
The .targets file was renamed as it must be named the same as the
.nuspec id otherwise the project the nuget package is installed into
won't run it.